AI campus assistant task 2/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.0.8
chromadb
python-dotenv

